requirements.txt
streamlit
requests
python-dotenv
nltk
pandas # Duygu analizi sonuçlarını veya veriyi işlemek için faydalı olabilir
python-dateutil # Tarih formatlama için